+++ src/session/coordinator.rs
@@ -263,7 +263,6 @@ impl SessionCoordinator {
             | SessionCommand::ResumeSession(_)
             | SessionCommand::NewSession
             | SessionCommand::ToggleMcpServer(_)
-            | SessionCommand::AnchoredToggle
             | SessionCommand::Interrupt => Ok(IdleDispatch::NotIdle),
         }
     }
@@ -778,7 +777,6 @@ impl SessionCoordinator {
             | SessionCommand::ResumeSession(_)
             | SessionCommand::NewSession
             | SessionCommand::ToggleMcpServer(_)
-            | SessionCommand::AnchoredToggle
             | SessionCommand::Interrupt => CommandOwnership::FrontendHosted,
         }
     }